기본 과제 수행 완료
