+++ src/refactoring/hooks/useCoupon.ts
@@ -2,5 +2,19 @@ import { Coupon } from "../../types.ts";
 import { useState } from "react";
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
-  return { coupons: [], addCoupon: () => undefined };
+  
+  const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
+  const addCoupon = (newCoupon) => {
+    setCoupons(prev => {
+      return [
+        ...prev,
+        newCoupon
+      ]
+    })
+  }
+  
+  return {
+    coupons,
+    addCoupon
+  };
 };
