+++ src/refactoring/hooks/useProduct.ts
@@ -2,5 +2,22 @@ import { useState } from 'react';
 import { Product } from '../../types.ts';
 
 export const useProducts = (initialProducts: Product[]) => {
-  return { products: [], updateProduct: () => undefined, addProduct: () => undefined };
+  const [products, setProducts] = useState<Product[]>(initialProducts);
+  
+  const updateProduct = (updatedProduct : Product) => {
+    setProducts(prevProducts =>
+      prevProducts.map(p => p.id === updatedProduct.id ? updatedProduct : p)
+    );
+  }
+  
+  const addProduct = (newProduct : Product) => {
+    setProducts(prev => {
+      return [
+        ...prev,
+        newProduct
+      ]
+    })
+  }
+  
+  return { products, updateProduct, addProduct };
 };
